Big refactoring part 1
In order to be able to also render emojis in media captions,
a ChatControlMessage now doesn't have different subclasses for each
message type, but rather a list of elements, which can be
thumbnail, text or smiley.

That way, it's easier to generate more generic (wow.) messages.

And, as a side-effect, also render smileys in media captions.